Add ResettableIterator
Adds ResettableIterator, an iterator type which caches the elements it
iterates over, so as to allow you to re-iterate over them at your
leisure.

Note that because this works by caching, it can consume large amounts of
memory if used on large iterators.

In general, I would suggest only using this on medium-sized or smaller
iterators you know you want to re-iterate over.