Add static function for constructing IMap more easily
Added a static function 'of' to IMap to allow you to more easily create
a map from a list of key-value pairs. However, misuse of this method can
result in getting ClassCastExceptions at some later point, because it
has to use Object var-args + an unsafe generic cast.

So, be careful, I suppose; and please make sure your argument types are
correct.