Add iterators which don't throw ConcurrentModificationException
This adds two iterators (one iterator, one ListIterator) which are
guaranteed not to throw a ConcurrentModificationException. The intended
use case is allowing you to iterate over a list and modify it at the
same time.

Note that this will still give odd behavior in the presence of actual
concurrency, but that is a deliberate tradeoff.