Add a whole bunch of Optics
This adds a whole bunch of types/functions related to optics. With this
batch, I've mainly gone for the concrete representation types, instead
of var Laarhoven or profunctor representations.

The concrete ones require less infrastructure, though they are not as
easy to compose

There's also a number of misc. things in here